# Optimize clipper: SIMD, reuse allocations, use slices