… benches/polygons.in
… benches/rectangles.dat