More to come, possibly

r? @glennw