@Manishearth 2015-10-03
Added some benchmarks based on real world data. `rectangles.dat` contains rectangle clipping data from various sites (500-1000 each), `polygons.in` contains all the polygon clipping data for the same.

Before:

```
running 2 tests
test bench_clip_poly ... bench:     116,924 ns/iter (+/- 2,995)
test bench_clip_rect ... bench:      78,551 ns/iter (+/- 1,992)
```

After:

```
test bench_clip_poly ... bench:     109,496 ns/iter (+/- 7,831)
test bench_clip_rect ... bench:      62,761 ns/iter (+/- 18,283)
```

The poly test has large variance, but multiple runs confirm an improvement.