@Manishearth 2015-10-03
With the shared clip buffers:

```
test bench_clip_poly ... bench:      99,667 ns/iter (+/- 6,136)
```

We still do some cloning. I think we can reduce some of that with some magic.