@Manishearth 2015-10-03
With the shared result buffer:

```
test bench_clip_poly ... bench:      94,444 ns/iter (+/- 5,819)
```