@Manishearth 2015-10-03
We should probably do some manual profiling of each step before merging. I'm not sure if all of these really end up optimizing it in actual usage. The numbers from testing on Servo looked promising, but they varied a lot.