@Manishearth 2015-10-09
Final numbers post rebase:

```
test bench_clip_poly ... bench:      88,945 ns/iter (+/- 11,819)
test bench_clip_rect ... bench:      57,166 ns/iter (+/- 3,722)
```


... I don't know how that happened. Perhaps SIMD got faster